+++ robolectric/src/main/java/org/robolectric/RobolectricTestRunner.java
@@ -48,7 +48,7 @@ import org.robolectric.internal.bytecode.ShadowMap;
 import org.robolectric.internal.bytecode.ShadowWrangler;
 import org.robolectric.internal.dependency.DependencyResolver;
 import org.robolectric.manifest.AndroidManifest;
-import org.robolectric.util.Injector;
+import org.robolectric.util.inject.Injector;
 import org.robolectric.util.PerfStatsCollector;
 import org.robolectric.util.ReflectionHelpers;
