+++ utils/src/main/java/org/robolectric/util/inject/Injector.java
@@ -1,4 +1,4 @@
-package org.robolectric.util;
+package org.robolectric.util.inject;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
