+++ utils/src/main/java/org/robolectric/util/inject/PluginFinder.java
@@ -1,4 +1,4 @@
-package org.robolectric.util;
+package org.robolectric.util.inject;
 
 import static java.util.Comparator.comparing;
