+++ utils/src/main/java/org/robolectric/util/inject/ServiceFinder.java
@@ -26,7 +26,7 @@
  * questions.
  */
 
-package org.robolectric.util;
+package org.robolectric.util.inject;
 
 import java.io.BufferedReader;
 import java.io.IOException;
