--- utils/src/test/java/org/robolectric/util/Thing.java
@@ -1,5 +1,0 @@
-package org.robolectric.util;
-
-interface Thing {
-
-}
