+++ utils/src/test/java/org/robolectric/util/inject/Thing.java
@@ -1,0 +1,5 @@
+package org.robolectric.util.inject;
+
+interface Thing {
+
+}
