--- utils/src/test/resources/META-INF/services/org.robolectric.util.Thing
@@ -1,1 +1,0 @@
-org.robolectric.util.InjectorTest$ThingFromServiceConfig
