+++ utils/src/test/resources/META-INF/services/org.robolectric.util.inject.Thing
@@ -1,0 +1,1 @@
+org.robolectric.util.inject.InjectorTest$ThingFromServiceConfig
